+++ src/math.ts
@@ -687,7 +687,7 @@ export function testLinePolygon(l: Line, p: Polygon): boolean {
 }
 
 export function testCirclePoint(c: Circle, p: Point): boolean {
-	return c.center.dist(p) < c.radius
+	return c.center.dist(p.toVec2()) < c.radius
 }
 
 export function testCircleCircle(c1: Circle, c2: Circle): boolean {